backend/requirements.txt
python-dotenv
flask
flask-cors
yoyo-migrations
flask-expects-json
mariadb
pymysql
cryptography